+++ app/src/main/java/com/pict/pbl/medswift/screens/user_auth/LoginActivity.kt
@@ -37,6 +37,7 @@ import androidx.compose.ui.unit.dp
 import androidx.lifecycle.MutableLiveData
 import com.pict.pbl.medswift.R
 import com.pict.pbl.medswift.login.LoginManager
+import com.pict.pbl.medswift.screens.HomeScreen
 import com.pict.pbl.medswift.screens.symptoms.SymptomsActivity
 import com.pict.pbl.medswift.ui.theme.MedSwiftTheme
 import com.pict.pbl.medswift.viewmodels.LoginViewModel
@@ -201,7 +202,7 @@ class LoginActivity : ComponentActivity() {
         Button(
             onClick = {
                 loginManager.loginUser( userEmail.value , userPassword.value ) {
-                    Intent( context , SymptomsActivity::class.java ).apply {
+                    Intent( context , HomeScreen::class.java ).apply {
                         startActivity( this )
                     }
                 }